parallelize using promise all
